Merge all dalex functionality in dalex-plots.py
This combines cateris_paribus and partial-dependence-profile (which used
to be 99% identical, so prime candidates for merging).

On top of that, a new functionality, plotting model-wise variable
importance has been added.

Also, the main function now has a more convenient configuration
structure, to make generating a single specific type of plot easy.

nonlinear-ANN still throws a NoneType exception. I'm leaving that to
@Garbarkozaurus. Hopefully it's an easy fix.